Adds initial docker config and scripts
There is still some work to be done around making this work end-to-end.
The main thing missing right now is copying the ssh private key for
bower to access the private repos.

I thought this would be a simple 'ADD' command in the Dockerfile:

  ADD ~/id_rsa /root/.ssh/

but it looks like this doesn't correctly expand. Using $HOME or similar
environment variables from the host also don't appear to work. I'm not
sure if there's another way around it, so it's something that could be
explored.

Worst-case, if this is only to be run by Jenkins, then we can just hard
code the Jenkins home path. I wanted to avoid this by dynamically
referencing the user's home directory so it could work across multiple
environments, but maybe that isn't possible.

I also looked into using Docker for running the environment while
developing, however changes don't appear to sync with Middleman.
Strangely enough, the files changes are correctly syncing into the
container, but they are not updated until Middleman is restarted. Not
sure why, as this wasn't a problem when running Sinatra.

This isn't a deal-breaker, as it's not terribly difficult to get
everything running locally without Docker